Update project structure and workflow
This commit includes substantial changes to the project structure and workflow. The CSS file is replaced with a SCSS version, and the primary .js file is updated. In addition, HTML structure modifications are made, a new clock.js and several .svg files are introduced. There's also an update in the project's building and deployment process via changes in 'vite.config.js', 'package.json' and introducing a new GitHub Actions workflow. Lastly, '.gitignore' and WebStorm configurations are added for better project setup.