Update the project from a clock to a minesweeper game
This commit transforms the project from a clock application to a fully functional minesweeper game. It includes the cleaning up of old clock-related files, adjustments to the main.js, replacing the SCSS file, addition of the svgs and edits to the index.html. There are significant changes within the project structure and workflow files, namely 'vite.config.js', 'package.json', and GitHub Actions workflow file. Other changes include an introduction of WebStorm configurations and edits to the .gitignore file.